kv_web/requirements.txt
Django==1.6.4
tipi==0.0.3
django-redis-cache==0.11.1
Pillow==2.4.0
